% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmriGraphList.R
\name{BrcFmriGraphList}
\alias{BrcFmriGraphList}
\title{\code{BrcFmriGraphList} Constructor}
\usage{
BrcFmriGraphList(graph.list, id, parcellation)
}
\arguments{
\item{graph.list}{a list of \code{BrcGraph} instances}

\item{id}{an identification string. It is not guaranteed to be
unique or nonempty---those depend on how the \code{BrcFmri} object
is constructed. Typically this field would be used to cross-reference
the MRI with a dataframe containing phenotype information.}

\item{parcellation}{a \code{BrcParcellation} instance}
}
\value{
a new \code{BrcFmriGraphList} instance
}
\description{
\code{BrcFmriGraphList} makes a new \code{BrcFmriGraphList} instance
}
\details{
A \code{BrcFmriGraphList} datatype contains all the information to 
represent a list of graphs within the MRI image. Specifically, it contains
a list of \code{BrcGraph} objects, stored in the \code{graph.list} element,
an \code{id} and the \code{BrcParcellation} object.

The number of nodes in each \code{BrcGraph} object in \code{graph.list}
must have the same number of unique parcellations in \code{parcellation}.
}

