% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcGraph.R
\name{BrcGraph}
\alias{BrcGraph}
\title{\code{BrcGraph} Constructor}
\usage{
BrcGraph(graph, type = NULL, parameter = NULL)
}
\arguments{
\item{graph}{an igraph object representing the graph}

\item{type}{a character representing the type of graph (statistically)}

\item{parameter}{a vector of numerics to specify how the graph was estimated}
}
\value{
a new \code{BrcGraph} instance
}
\description{
\code{BrcGraph} makes a new \code{BrcGraph} instance
}
\details{
A \code{BrcGraph} datatype contains all the information on the graph
as an \code{igraph} representation, the type of graph it statistically
represents, and the parameters needed when estimating the graph.

Because \code{BrcGraph} is a dataype, not an object there are no accessor 
functions. You can get its components directly with the \code{$} operator.
In particular, you can use all of functions in \code{igraph} to interact
with the \code{graph} element in \code{BrcGraph}
}

