% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGraph.BrcFmri.R
\name{BrcGraphTypes}
\alias{BrcGraphTypes}
\title{Types of Graphs to be used in \code{estimateGraph}}
\usage{
BrcGraphTypes()
}
\value{
a vector of available graphs, to be passed into "method" of
\code{estimateGraph}.
}
\description{
Types of Graphs to be used in \code{estimateGraph}
}

