% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGraph.BrcFmri.R
\name{estimateGraph}
\alias{estimateGraph}
\title{Estimate Connectome (Graph) from BrcFmri data}
\usage{
estimateGraph(x, method = "correlation", ...)
}
\arguments{
\item{x}{the \code{BrcFmri} object}

\item{method}{the method to use, must be one of the methods shown in
\code{BrcGraphTypes()}. "correlation" is the default.}

\item{...}{additional parameters to be used when estimating the graph}
}
\value{
a new \code{BrcFmriGraphList} instance
}
\description{
The implementation for "partialCorrelation.glasso" and 
"partialCorrleation.neighborhoodselect" is based on the \code{HUGE} package.
The graphs output will have nodes, one for each unique parcel in 
x. The \code{BrcFmriGraphList} will contain a list of BrcGraphs, each 
corresponding to a different set of tuning parameters.
}

