% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcFmriGraphList.R
\name{isValid.BrcFmriGraphList}
\alias{isValid.BrcFmriGraphList}
\title{Checking \code{BrcFmriGraphList} instance validity.}
\usage{
isValid.BrcFmriGraphList(obj)
}
\arguments{
\item{obj}{the \code{BrcFmriGraphList} object}
}
\value{
TRUE if valid
}
\description{
\code{isValid} method for class "\code{BrcFmriGraphList}.
}
\details{
Fails noisily with a stop message if the \code{BrcFmriGraphList}
instance is invalid. Otherwise, returns TRUE.
}

