% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcGraph.R
\name{isValid.BrcGraph}
\alias{isValid.BrcGraph}
\title{Checking BrcGraph instance validity.}
\usage{
isValid.BrcGraph(obj)
}
\arguments{
\item{obj}{the \code{BrcGraph} object}
}
\value{
TRUE if valid
}
\description{
\code{isValid} method for class \code{BrcGraph}.
}
\details{
Fails noisily with a stop message if the \code{BrcGraph}
instance is invalid. Otherwise, returns TRUE.
}

