% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericFunctions.R
\name{isValid}
\alias{isValid}
\title{Checks S3 object validity}
\usage{
isValid(obj)
}
\arguments{
\item{obj}{The object to check}
}
\value{
void
}
\description{
Uses the generic function in \code{brcbase}. This
is here only to get the \code{brcstat} package to
compile.
}

