% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrcGraph.R
\name{makeGraphDefault}
\alias{makeGraphDefault}
\title{Construct an \code{igraph} object from a matrix of nodes}
\usage{
makeGraphDefault(edges, n = max(edges), ...)
}
\arguments{
\item{edges}{a matrix or vector of numerics (positive integers) representing
the edges between nodes}

\item{n}{the total number of nodes}

\item{...}{additional parameters to be passed into igraph::make_graph}
}
\value{
a new \code{igraph} instance
}
\description{
\code{makeGraphDefault} makes a new \code{igraph} instance.
}
\details{
The \code{edges} can represent a matrix or vector. If there are 
isolated edges, be sure to use the \code{n} parameter. The default
value of \code{n} is the maximum value in \code{edges}. The unique
values in \code{edges} do not need to have unique, consecutive values.
If a matrix is passed in for \code{edges}, the code will implicitly
convert the matrix into \code{as.numeric(edges)}, and each consecutive
pair of values will represent an edge.

The graph outputted is simplified, meaning there are no self-edges
(i.e., loops, or edges that start and end at the same node) or
duplicated edges. It is also undirected.
}

